// src/app/court/models/VideoCallUser.ts
export class VideoCallUser{
    displayName: string;
    userAffidavitId: number;
    currentStatus: string;
    price: number;
    templateName: string;
    affidavitUserEmail:string;
    userAffidavitCustomName:string;
    isUploaded:Boolean;
    deponentEmail:String;
    deponentName:String;
    requesterName:string;
    deponentId;
    selected: boolean;
    isCaseRelated:Boolean;
    signatureStatus: string;
    signatureDocumentId: string;
    userSigned: boolean;
}